DrawTextW 関数の第3引数の型 int に合わせて呼び出すように記述調整

+++ sakura_core/dlg/CDlgAbout.cpp
@@ -539,7 +539,7 @@ bool CUrlWnd::OnSetText( _In_opt_z_ LPCWSTR pchText, _In_opt_ size_t cchText ) c
 	// DrawText関数を使ってサイズを計測する
 	// ※この処理は実際には描かない
 	CMyRect rcText;
-	int retDrawText = ::DrawText( hDC, pchText, cchText, &rcText, DT_CALCRECT );
+	int retDrawText = ::DrawText( hDC, pchText, static_cast<int>(cchText), &rcText, DT_CALCRECT );
 
 	// DCの後始末
 	::SelectObject( hDC, hObj );
