+++ sakura_core/uiparts/CMenuDrawer.cpp
@@ -1163,7 +1163,7 @@ void CMenuDrawer::DrawItem( DRAWITEMSTRUCT* lpdis )
 	::DrawText(
 		hdc,
 		pszItemStr,
-		j,
+		static_cast<int>(j),
 		&rcText,
 		DT_LEFT | DT_VCENTER | DT_SINGLELINE
 	);
@@ -1172,7 +1172,7 @@ void CMenuDrawer::DrawItem( DRAWITEMSTRUCT* lpdis )
 		::DrawText(
 			hdc,
 			&pszItemStr[j + 1],
-			nItemStrLen - ( j + 1 ),
+			static_cast<int>(nItemStrLen - ( j + 1 )),
 			&rcText,
 			DT_RIGHT | DT_VCENTER | DT_SINGLELINE
 		);
